Fixed bug in cub reader for meshes with non-contiguous element numbering.
Changed algorithm used to restore geometric topology from elements, at the cost of memory.  For a 
large geometry with relatively few elements the cost is maybe 200MB; could be higher for a large mesh,
so I'll eventually change this to be controllable.

Fixed a few compile warnings in WriteSLAC.

Reading in a large cub file (> 5k volumes, > 20k surfaces, ~100k hexes) takes < 20 seconds, unoptimized.


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@1478 6162379e-bd28-0410-9a7d-b7f4dcfcad3c

+++ Tqdcfr.hpp
@@ -273,8 +273,8 @@ public:
   FileTOC fileTOC;
   ModelEntry *modelEntries;
   MetaDataContainer modelMetaData;
-  int currNodeIdOffset;
-  int currElementIdOffset[MBMAXTYPE];
+  long currNodeIdOffset;
+  long currElementIdOffset[MBMAXTYPE];
   MBTag globalIdTag, cubIdTag, geomTag, uniqueIdTag, blockTag, nsTag, ssTag,
     attribVectorTag, entityNameTag, categoryTag;
   std::map<int, MBEntityHandle> uidSetMap;
